Fabrix: Linux Integration            - Added the fabric module and icon to the proper CMakeLists.txt            - Fixed Panel to support it being added/removed without being destroyed by gtk
[TESTS]
- Did a regression testing with all the features just as donw in windows/mac and everything worked as expected.

+++ backend/wbprivate/workbench/home_screen.cpp
@@ -1476,7 +1476,7 @@ public:
 
           // Creates the legend to be displayed on the filter icon
           if (groups.size() > 2)
-            tile_groups = base::strfmt("%s and %d others", groups[0].c_str(), groups.size() - 1);
+            tile_groups = base::strfmt("%s and %li others", groups[0].c_str(), (long)groups.size() - 1);
           else
             tile_groups = ha_filter;
         }
